chore (OAuthService) : 회원가입시 member에들어오는 null값 수정

+++ src/main/java/chzzk/grassdiary/domain/member/Member.java
@@ -45,7 +45,7 @@ public class Member extends BaseTimeEntity {
     @Lob
     private String picture; // auth
 
-    private Long grassCount;
+    private int grassCount;
 
     @ColumnDefault("0")
     private int rewardPoint;
